+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+### Added
+
+- Added documentation about `timeout` and `timeoutlen`
+
 ## 0.2 - 2021-02-17
 
 ### Added