+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 
 ### Added
 
+- Make FontFamily configurable
 - Added documentation about `timeout` and `timeoutlen`
 
 ## 0.2 - 2021-02-17