+++ CHANGELOG.md
@@ -8,6 +8,10 @@
 - Make FontFamily and FontSize configurable
 - Added documentation about `timeout` and `timeoutlen`
 
+### Changed
+
+- Arguments for VIM motions (e.g. f, t, F, T) do not trigger the popup anymore
+
 ## 0.2 - 2021-02-17
 
 ### Added